Meio finalizado

+++ ultranet/ultranet.ino
@@ -27,7 +27,7 @@ void loop()
 {
   //Le as informacoes do sensor, em cm e pol
   float cmMsec, inMsec;
-  long microsec = ultrasonic.timing();
+  float microsec = ultrasonic.timing();
   cmMsec = ultrasonic.convert(microsec, Ultrasonic::CM);
   inMsec = ultrasonic.convert(microsec, Ultrasonic::IN);
   //Exibe informacoes no serial monitor
